+++ apps/App.js
@@ -1,0 +1,2 @@
+import App from './apps/mobile/src/App';
+export default App;