+++ apps/mobile/index.ts
@@ -1,5 +1,4 @@
 import { registerRootComponent } from 'expo';
-
-import App from './App';
+import App from './App'; // App.tsx aynı klasörde olduğu için direkt böyle çağırıyoruz
 
 registerRootComponent(App);